Revert "feat: get texas and western hydro v2"
