+++ prereise/gather/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['demanddata', 'solardata', 'winddata', 'hydrodata','get_monthly_net_generation','trim_eia_form_923']
+__all__ = ['demanddata', 'solardata', 'winddata', 'hydrodata']
