--- prereise/gather/hydrodata/data/cf.csv
@@ -1,36 +1,0 @@
-01/15/2015, 0.407
-02/15/2015, 0.414
-03/15/2015, 0.408
-04/15/2015, 0.394
-05/15/2015, 0.339
-06/15/2015, 0.358
-07/15/2015, 0.358
-08/15/2015, 0.325
-09/15/2015, 0.283
-10/15/2015, 0.283
-11/15/2015, 0.338
-12/15/2015, 0.394
-01/15/2016, 0.436
-02/15/2016, 0.438
-03/15/2016, 0.459
-04/15/2016, 0.446
-05/15/2016, 0.428
-06/15/2016, 0.406
-07/15/2016, 0.361
-08/15/2016, 0.330
-09/15/2016, 0.286
-10/15/2016, 0.293
-11/15/2016, 0.328
-12/15/2016, 0.379
-01/15/2017, 0.454
-02/15/2017, 0.441
-03/15/2017, 0.491
-04/15/2017, 0.511
-05/15/2017, 0.547
-06/15/2017, 0.527
-07/15/2017, 0.451
-08/15/2017, 0.373
-09/15/2017, 0.334
-10/15/2017, 0.310
-11/15/2017, 0.360
-12/15/2017, 0.377
