--- prereise/gather/tests/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ['mock_generation_data_frame','test_get_monthly_net_generation']
