+++ setup.py
@@ -9,8 +9,5 @@ setup(name='prereise',
       packages=find_packages(),
       package_data={'prereise': ['call/add_path.m',
                                  'gather/winddata/IECPowerCurves.csv',
-                                 'gather/hydrodata/data/cf.csv',
-                                 'gather/data/EIA923_Schedules_2_3_4_5_M_12_2016_Final_Revision.xlsx',
-                                 'gather/hydrodata/data/hydro_gmt.csv',
-                                 'gather/hydrodata/data/texas_hydro_gmt.csv']},
+                                 'gather/hydrodata/eia/cf.csv']},
       zip_safe=False)
